Add mucjid in the Presence "primary key", an actual key for PostgreSQL and an unique hash column for MySQL (3072 key length limit)

+++ database/migrations/20250109192438_add_mucjid_presences_primary_key.php
@@ -6,29 +6,32 @@ use Illuminate\Database\Capsule\Manager as DB;
 
 class AddMucjidPresencesPrimaryKey extends Migration
 {
+    /**
+     * This migration is failing on MySQL (3072 key length limit)
+     * We pass it and create a hash of the needed columns the next migration
+     */
+
     public function up()
     {
-        $this->disableForeignKeyCheck();
+        if ($this->schema->getConnection()->getDriverName() == 'pgsql') {
+            DB::table('presences')->delete();
 
-        DB::table('presences')->delete();
-
-        $this->schema->table('presences', function (Blueprint $table) {
-            $table->dropPrimary(['session_id', 'jid', 'resource']);
-            $table->primary(['session_id', 'jid', 'mucjid', 'resource']);
-        });
-
-        $this->enableForeignKeyCheck();
+            $this->schema->table('presences', function (Blueprint $table) {
+                $table->dropPrimary(['session_id', 'jid', 'resource']);
+                $table->primary(['session_id', 'jid', 'mucjid', 'resource']);
+            });
+        }
     }
 
     public function down()
     {
-        $this->disableForeignKeyCheck();
+        if ($this->schema->getConnection()->getDriverName() == 'pgsql') {
+            DB::table('presences')->delete();
 
-        $this->schema->table('presences', function (Blueprint $table) {
-            $table->dropPrimary(['session_id', 'jid', 'mucjid', 'resource']);
-            $table->primary(['session_id', 'jid', 'resource']);
-        });
-
-        $this->enableForeignKeyCheck();
+            $this->schema->table('presences', function (Blueprint $table) {
+                $table->dropPrimary(['session_id', 'jid', 'mucjid', 'resource']);
+                $table->primary(['session_id', 'jid', 'resource']);
+            });
+        }
     }
 }
